LayoutTests/editing/pasteboard/paste-image-does-not-reveal-file-url-expected.txt

Tests that pasting images do not reveal its file URL. To manually test, copy the image below in the context menu and paste into the red box.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.clipboardData.types.includes("Files") is true
PASS event.clipboardData.getData("url") is ""
PASS event.clipboardData.getData("text/plain") is ""
PASS event.clipboardData.getData("text/uri-list") is ""
PASS successfullyParsed is true

TEST COMPLETE
